requirements.txt
puremagic
androguard==3.4.0a1
apkid
rich
tqdm
colorama
oletools
pefile
quark-engine==21.8.1
pyaxmlparser
yara-python
prompt_toolkit==3.0.19
frida
lief